generalized train_predictor_trial.bash
Generalized train_predictor_trial.bash file: Added “struct” (to
designate minimized or unminimized structure) and “r_cut” (to designate
applied distance cut-off) options to be supplied from the command line

When I initially performed the feature extraction, I added a
descriptive line to some of the output files. Since every output file
did not contain this line, I removed it from the files that did so that
text file loading would be easier in the future.